Attempting to fix failing unit tests around RemoveWord (they pass in the Realm.Library unit tests, but fail in Smaug CS) Moved load and save watchlist into its own loader

+++ trunk/SmaugCS/Objects/WatchData.cs
@@ -2,7 +2,7 @@
 {
     public class WatchData
     {
-        public short ImmortalLevel { get; set; }
+        public int ImmortalLevel { get; set; }
         public string ImmortalName { get; set; }
         public string TargetName { get; set; }
         public string PlayerSite { get; set; }
